# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.ComponentDeploymentListResponse:
    properties:
      deployments:
        items:
          $ref: '#/definitions/service.ComponentDeploymentResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  service.ComponentDeploymentResponse:
    properties:
      build_properties:
        type: object
      build_time:
        type: string
      component_id:
        type: string
      created_at:
        type: string
      deployed_at:
        type: string
      git_commit_id:
        type: string
      git_commit_time:
        type: string
      git_properties:
        type: object
      id:
        type: string
      is_active:
        type: boolean
      landscape_id:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  service.CreateComponentDeploymentRequest:
    properties:
      build_properties:
        type: object
      build_time:
        type: string
      component_id:
        type: string
      deployed_at:
        type: string
      git_commit_id:
        type: string
      git_commit_time:
        type: string
      git_properties:
        type: object
      is_active:
        type: boolean
      landscape_id:
        type: string
      version:
        type: string
    required:
    - component_id
    - landscape_id
    type: object
...
  service.UpdateComponentDeploymentRequest:
    properties:
      build_properties:
        type: object
      build_time:
        type: string
      deployed_at:
        type: string
      git_commit_id:
        type: string
      git_commit_time:
        type: string
      git_properties:
        type: object
      is_active:
        type: boolean
      version:
        type: string
    type: object
...
host: localhost:7008
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
...
  /component-deployments:
    get:
      consumes:
      - application/json
      description: Get component deployments filtered by component or landscape with
        pagination
      parameters:
      - description: Component ID (UUID) to filter deployments
        in: query
        name: component_id
        type: string
      - description: Landscape ID (UUID) to filter deployments
        in: query
        name: landscape_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved component deployments
          schema:
            $ref: '#/definitions/service.ComponentDeploymentListResponse'
        "400":
          description: Invalid parameters - component_id or landscape_id required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List component deployments
      tags:
      - component-deployments
    post:
      consumes:
      - application/json
      description: Deploy a component to a specific landscape
      parameters:
      - description: Component deployment data
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/service.CreateComponentDeploymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created component deployment
          schema:
            $ref: '#/definitions/service.ComponentDeploymentResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Component or landscape not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Component deployment already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new component deployment
      tags:
      - component-deployments
  /component-deployments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a component deployment by ID
      parameters:
      - description: Component deployment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted component deployment
        "400":
          description: Invalid component deployment ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Component deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete component deployment
      tags:
      - component-deployments
    get:
      consumes:
      - application/json
      description: Get a specific component deployment by its UUID
      parameters:
      - description: Component deployment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved component deployment
          schema:
            $ref: '#/definitions/service.ComponentDeploymentResponse'
        "400":
          description: Invalid component deployment ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Component deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get component deployment by ID
      tags:
      - component-deployments
    put:
      consumes:
      - application/json
      description: Update an existing component deployment by ID
      parameters:
      - description: Component deployment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated component deployment data
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/service.UpdateComponentDeploymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated component deployment
          schema:
            $ref: '#/definitions/service.ComponentDeploymentResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Component deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update component deployment
      tags:
      - component-deployments
  /component-deployments/{id}/details:
    get:
      consumes:
      - application/json
      description: Get a component deployment with all its related data
      parameters:
      - description: Component deployment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved component deployment with details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid component deployment ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Component deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get component deployment with full details
      tags:
      - component-deployments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"